# contracts/openapi/identity-service.v1.yaml
openapi: 3.1.0
info:
  title: AgriCore Identity Service API
  version: 1.4.0
  description: |
    Authentication, users, RBAC, and migration-owned permission policy.
    Browser clients should prefer `/api/v1/auth/web/*` so refresh credentials stay
    in an HttpOnly cookie. Non-browser clients continue to use JSON refresh tokens.
    Pre-release compatibility note: version 1.4.0 intentionally removes runtime
    permission creation and makes permission-policy replacement versioned and audited.
servers:
  - url: http://localhost:8081
paths:
...
  /api/v1/admin/users:
    get:
      summary: List users for administration
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Paged users ordered by creation time descending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/users/{userId}/roles:
    patch:
      summary: Replace a user's role set
      operationId: updateUserRoles
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRolesRequest'
      responses:
        '200':
          description: Roles replaced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
...
components:
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 2147483646
        default: 0
    Size:
      name: size
      in: query
      required: false
      description: Values above 100 are capped to 100 by the service.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 2147483647
        default: 20
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: The request is malformed, invalid, or fails validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Bearer credential missing, expired, or invalid. The security filter chain may return this response without a body.
    Forbidden:
      description: Authenticated, but missing the required role or permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: The requested resource does not exist or is not visible to the caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    UnsupportedMediaType:
      description: The request Content-Type is not supported.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    UserResponse:
      type: object
      required: [id, email, fullName, status, roles, permissions, lastLoginAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email, maxLength: 320 }
        fullName: { type: string, maxLength: 200 }
        status: { type: string, enum: [ACTIVE, LOCKED, DISABLED] }
        roles:
          type: array
          maxItems: 7
          uniqueItems: true
          description: |
            Sorted role codes. Auth token responses match the issued JWT snapshot;
            `/users/me` reflects the authenticated bearer token snapshot.
          items:
            $ref: '#/components/schemas/RoleCode'
        permissions:
          type: array
          maxItems: 100
          uniqueItems: true
          description: |
            Sorted effective canonical permissions. Auth token responses match the issued JWT
            snapshot; `/users/me` reflects the authenticated bearer token snapshot.
          items:
            $ref: '#/components/schemas/PermissionCode'
        lastLoginAt:
          type: [string, 'null']
          format: date-time
        createdAt: { type: string, format: date-time }
    RoleCode:
      type: string
      enum:
        - SYSTEM_ADMIN
        - FARM_MANAGER
        - AGRONOMIST
        - FIELD_WORKER
        - WAREHOUSE_MANAGER
        - SALES_STAFF
        - AUDITOR
    UpdateUserRolesRequest:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          minItems: 1
          maxItems: 7
          uniqueItems: true
          items:
            $ref: '#/components/schemas/RoleCode'
    PermissionCode:
      type: string
      minLength: 3
      maxLength: 64
      pattern: '^[A-Z][A-Z0-9_]{2,63}$'
...
    UserPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, first, last]
      properties:
        content:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/UserResponse'
        page: { type: integer, format: int32, minimum: 0, maximum: 2147483646 }
        size: { type: integer, format: int32, minimum: 1, maximum: 100 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, format: int32, minimum: 0 }
        first: { type: boolean }
        last: { type: boolean }
    ApiError:
      type: object
      description: |
        Uniform error body for every AgriCore service, defined by ApiError in common-lib.
        `code` is the stable machine-readable discriminator and is part of the public contract.
        Null fields are omitted, so `violations`, `details`, and `traceId` are absent unless set.
      required: [timestamp, status, error, code, message, path]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string, description: HTTP reason phrase }
        code: { type: string, example: DUPLICATE_RESOURCE }
        message: { type: string }
        path: { type: string }
        traceId: { type: string }
        violations:
          type: array
          description: Present only when code is VALIDATION_FAILED.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
        details:
          type: object
          additionalProperties: true
...